Move notification permission logic to its own file
